refactor(router): extract type resolvers to new file
