+++ packages/router/src/type-resolvers.ts
@@ -1,0 +1,73 @@
+import { Constructable } from '@aurelia/kernel';
+import { CustomElement, ICustomElementType } from '@aurelia/runtime';
+import { ComponentHandle, IRouteableComponent, IRouteableComponentType, IViewportInstruction, NavigationInstruction, ViewportHandle } from './interfaces';
+import { IRouter } from './router';
+import { Viewport } from './viewport';
+import { ViewportInstruction } from './viewport-instruction';
+
+export const ComponentHandleResolver = {
+  isName: function <T>(component: T & ComponentHandle<Constructable>): component is T & ComponentHandle<Constructable> {
+    return typeof component === 'string';
+  },
+  isType: function <T>(component: T & ComponentHandle<Constructable>): component is T & ComponentHandle<Constructable> {
+    return CustomElement.isType(component);
+  },
+  getName: function <T>(component: T & ComponentHandle<Constructable>): string {
+    if (ComponentHandleResolver.isName(component)) {
+      return component as string;
+    } else if (ComponentHandleResolver.isType(component)) {
+      return (component as ICustomElementType).description.name;
+    } else {
+      return ((component as IRouteableComponent).constructor as ICustomElementType).description.name;
+    }
+  },
+  getType: function <T extends Constructable>(component: T & ComponentHandle<Constructable>): ICustomElementType<T> {
+    if (ComponentHandleResolver.isName(component)) {
+      return null;
+    } else if (ComponentHandleResolver.isType(component)) {
+      return component;
+    }
+    // TODO: Fix resolve for instance
+    // else {
+    //   return (component as ICustomElementType).constructor;
+    // }
+  },
+};
+
+export const ViewportHandleResolver = {
+  isName: function <T>(viewport: T & ViewportHandle): viewport is T & ViewportHandle {
+    return typeof viewport === 'string';
+  },
+  isInstance: function <T>(viewport: T & ViewportHandle): viewport is T & ViewportHandle {
+    return viewport instanceof Viewport;
+  },
+  getName: function <T>(viewport: T & ViewportHandle): string {
+    if (ViewportHandleResolver.isName(viewport)) {
+      return viewport as string;
+    } else {
+      return viewport ? (viewport as Viewport).name : null;
+    }
+  }
+};
+
+export const NavigationInstructionResolver = {
+  toViewportInstructions: function <C extends Constructable>(router: IRouter, navigationInstructions: NavigationInstruction | NavigationInstruction[]): ViewportInstruction[] {
+    if (!Array.isArray(navigationInstructions)) {
+      return NavigationInstructionResolver.toViewportInstructions(router, [navigationInstructions]);
+    }
+    const instructions: ViewportInstruction[] = [];
+    for (const instruction of navigationInstructions) {
+      if (typeof instruction === 'string') {
+        instructions.push(router.instructionResolver.parseViewportInstruction(instruction));
+      } else if (instruction as ViewportInstruction instanceof ViewportInstruction) {
+        instructions.push(instruction as ViewportInstruction);
+      } else if (instruction['component']) {
+        const viewportComponent = instruction as IViewportInstruction<C>;
+        instructions.push(new ViewportInstruction(viewportComponent.component, viewportComponent.viewport, viewportComponent.parameters));
+      } else {
+        instructions.push(new ViewportInstruction(instruction as IRouteableComponentType<C>));
+      }
+    }
+    return instructions;
+  },
+};
